requirements.txt
mkdocs-eds @ git+https://github.com/percevalw/mkdocs-eds.git@main#egg=mkdocs-eds
pret-joy
pybtex<=0.24.0